combine a mul-sum pair to a mma op
