+++ CMakeLists.txt
@@ -402,6 +402,7 @@ list(APPEND JIT_TEST_SRCS
   ${NVFUSER_ROOT}/test/test_gpu_indexing_ops.cpp
   ${NVFUSER_ROOT}/test/test_gpu_indexing.cpp
   ${NVFUSER_ROOT}/test/test_combined_inner_outer_reduction.cpp
+  ${NVFUSER_ROOT}/test/test_combine_mul_sum.cpp
   ${NVFUSER_ROOT}/test/test_optimization_pass.cpp
   ${NVFUSER_ROOT}/test/test_pipeline.cpp
   ${NVFUSER_ROOT}/test/test_serial_gridreduce.cpp
